+++ spp_dci/__manifest__.py
@@ -2,7 +2,7 @@
     "name": "OpenSPP DCI Core",
     "summary": "Core DCI (Digital Convergence Initiative) API components",
     "category": "OpenSPP/Integration",
-    "version": "19.0.2.0.1",
+    "version": "19.0.2.0.2",
     "author": "OpenSPP.org",
     "website": "https://github.com/OpenSPP/OpenSPP2",
     "license": "LGPL-3",